TClock Light kt090704
* merge some patches from tclocklight-050716.
* add config.h to support enabling/disabling the features.
* an additional modification for x64.

+++ source/exe/alarm.c
@@ -180,7 +180,7 @@ void OnTimerAlarm(HWND hwnd, const SYSTEMTIME* st, int reason)
 		}
 		if(reason == 0 && pAS->bResumeTimer)
 		{
-			if(GetTickCount() - resumeTick > pAS->nResumeDelay * 1000)
+			if(GetTickCount() - resumeTick > (DWORD)(pAS->nResumeDelay * 1000))
 			{
 				bexec = TRUE;
 				pAS->bResumeTimer = FALSE;
